free disk space in runner image
The integration tests in Github actions were failing due to insufficient
disk space available in the runner image. These runners come preloaded
with a substantial number of libraries and tools meant to accommodate a
vast, diverse set of projects.

Remove the Android libraries from our runner image prior to executing
the integration tests. Android is the largest single consumer of disk
space and it is almost certainly something that Hermeto will never use.

Signed-off-by: Taylor Madore <tmadore@redhat.com>