tests: integration: pip: Disable the Rust extensions test
The test has been changed upstream and we'd have to backport Rust
specific stuff to the 0.28 branch. We're primarily motivated by bumping
the supported Go version, so skip this problematic test for now. We can
re-consider enabling it again if we're to backport Rust/Cargo stuff to
another 0.28 patch release.

Signed-off-by: Erik Skultety <eskultet@redhat.com>