feat: Add sourceInfo to SPDX SBOM
This commit adds sourceInfo to SPDX SBOM. This field will
be used to track source URL actually used for a package unlike a
source URL reported in a PURL. This is necessary when a package is
downloaded through a proxy. To accommodate for cases when more than
one proxy URL could be specified  (for example, for golang) individual
URLs should be concatenated with a semicolon.

Signed-off-by: Alexey Ovchinnikov <aovchinn@redhat.com>